+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/LineGraph/BasicLineGraphs/DashedLine.jsx
@@ -1,14 +1,28 @@
 import React from 'react';
-import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/dashedLineData';
+import dashedLineData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/dashedLineData';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(lineDefault('#dashedLine'));
-const dataset = utils.deepClone(data);
+const dataset = utils.deepClone(dashedLineData);
 
-export default () => <LineGraph graphID="dashedLine" graphConfig={graphConfig} dataset={dataset} />;
+const LineExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <div id="dashedLine" />
+  );
+};
+
+export default LineExample;