+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/LineGraph/BasicLineGraphs/TimeseriesEventline.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
-import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
@@ -26,11 +27,22 @@ graphConfig.eventline = [
     value: new Date(2016, 0, 1, 12).toISOString(),
   },
 ];
-const dataset = utils.deepClone(data);
+const dataset = utils.deepClone(exampleData);
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="timeseriesEventLine" graphConfig={graphConfig} dataset={dataset} />
-  </React.Fragment>
-);
+const LineExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="timeseriesEventLine" />
+    </React.Fragment>
+  );
+};
+
+export default LineExample;
+