+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/LineGraph/Criticality/MultiLine.jsx
@@ -1,16 +1,17 @@
 import React from 'react';
-import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
-
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(lineDefault('#criticalityMultiLine'));
 
-const contentData = utils.deepClone(data);
+const contentData = utils.deepClone(exampleData);
 contentData[0].values[0].isCritical = true;
 contentData[0].values[5].isCritical = true;
 contentData[0].values[10].isCritical = true;
@@ -18,10 +19,20 @@ contentData[0].values[10].isCritical = true;
 contentData[1].values[6].isCritical = true;
 contentData[1].values[11].isCritical = true;
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="criticalityMultiLine" graphConfig={graphConfig} dataset={contentData} />
-  </React.Fragment>
-);
+const LineExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    contentData.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="criticalityMultiLine" />
+    </React.Fragment>
+  );
+};
+
+export default LineExample;