+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/LineGraph/Regions/Line/NoLowerBound.jsx
@@ -1,25 +1,36 @@
 import React from 'react';
-import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
-
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(lineDefault('#noLowerBound'));
-const contentData = utils.deepClone(data);
+const contentData = utils.deepClone(exampleData);
 contentData[0].regions = [
   {
     end: 10,
   },
 ];
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="noLowerBound" graphConfig={graphConfig} dataset={contentData} />
-  </React.Fragment>
-);
+const LineExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    contentData.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="noLowerBound" />
+    </React.Fragment>
+  );
+};
+
+export default LineExample;