+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/BasicLineGraphs/TimeseriesWithDateline.test.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
-import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
 import {
   loadDatelinePopup,
 } from '@cerner/terra-graphs-docs/lib/example-datasets/popup';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
 
 const graphConfig = utils.deepClone(lineTimesries('#timeseriesDateLine'));
 graphConfig.dateline = [
@@ -22,11 +23,22 @@ graphConfig.dateline = [
     value: new Date(2016, 0, 1, 6).toISOString(),
   },
 ];
-const dataset = utils.deepClone(data);
+const dataset = utils.deepClone(exampleData);
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="timeseriesDateLine" graphConfig={graphConfig} dataset={dataset} />
-  </React.Fragment>
-);
+const LineExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="timeseriesDateLine" />
+    </React.Fragment>
+  );
+};
+
+export default LineExample;
+